+++ src/main/java/com/evacipated/cardcrawl/mod/stslib/Keyword.java
@@ -1,0 +1,8 @@
+package com.evacipated.cardcrawl.mod.stslib;
+
+public class Keyword
+{
+    public String PROPER_NAME;
+    public String[] NAMES;
+    public String DESCRIPTION;
+}