+++ src/main/java/com/evacipated/cardcrawl/mod/stslib/cards/interfaces/StartupCard.java
@@ -1,0 +1,6 @@
+package com.evacipated.cardcrawl.mod.stslib.cards.interfaces;
+
+public interface StartupCard
+{
+    void atBattleStartPreDraw();
+}